+++ pkg/network/java/testdata/netty/pom.xml
